Module Bench_common.Generate_trees

Parameters

module Store : Irmin.Generic_key.KV with type Schema.Contents.t = bytes

Signature

val add_chain_trees : int -> int -> Store.tree -> Store.tree Lwt.t

add_chain_trees depth nb tree adds nb random contents to tree, depthwise.

val add_large_trees : int -> int -> Store.tree -> Store.tree Lwt.t

add_large_trees width nb tree adds nb random contents to tree, breadthwise.