Module Irmin_traces.Trace_definitions

Traces file format definitions.

This file is meant to be used from Tezos. OCaml version 4.09 and the 32bit architecture should be supported.

Traces Workflow

A Tezos node (may) output a Raw_replayable_trace file. Such a trace should be postprocessed to create a Replayable_trace.

A Tezos node (may) output a Stat_trace file.

trace_stats.exe summarise takes a Stat_trace file and summarises it to a stat_summary.json file.

A series of python script take a stat_summary.json file and produce plots (e.g. png files).

tree.exe takes a Replayable_trace file, internally produces a Stat_trace file and yields it to Trace_stat_summary to produce a stat_summary.json file.

module Replayable_trace : sig ... end

Replayable_trace, a trace of Tezos's interactions with Irmin.

module Stat_trace : sig ... end

Trace of a tezos node run, or a replay run.