Module Trace_stat_summary.Span

A stat trace can be chunked into blocks. A blocks is made of 2 phases, first the buildup and then the commit.

The duration of a buildup can be split into 2 parts: 1. the time spend in each operation and 2. the sum of the time spent between, before and after all operations. The first being the seen buildup and the second the unseen buildup.

The total duration of a block is the sum of the durations of the commit, the seen buildup and the unseen buildup.

Caveat: There isn't a one to one correspondance between summary blocks and Tezos' blocks. A Tezos block is associated to a commit, but a commit is not necessarily associated to a Tezos block. There are ~50 more commits than Tezos blocks up to the Edo protocol.

module Key : sig ... end

module Val : sig ... end

module Map : sig ... end

type map = Val.t Map.t

val map_t : map Irmin.Type.t