Module Span.Val

type t = {

count : Vs.t;

cumu_count : Vs.t;

duration : Vs.t;

duration_log_scale : Vs.t;

cumu_duration : Vs.t;

}

The count variable is the number of occurences of a span per block and cumu_count is the cumulative from the beginning.

The duration variable is the length of a span occurence and cumu_duration is the cumulative from the beginning. The x axis for the evolution is the number of blocks.

The count for commit and unseen is trivialy equal to 1. The same is almost true for checkout too.

val t : t Repr.t