Module Irmin_traces.Trace_stat_summary_conf

Summary configuration.

This file is NOT meant to be used from Tezos, as opposed to some other "trace_*" files.

val histo_bin_count : int

val curves_sample_count : int

val moving_average_half_life_ratio : float

Parameter to control the width of the moving average window.

This width is expressed as a fraction of the width of the final plot images, (i.e. the number of played blocks). This implies that the amount of smoothing remains visually the same, no matter curves_sample_count and the number of blocks in the stat trace.

Justification of the current value:

The 2nd commit of the full replay trace is a massive one, it contains ~1000x more operations than an average one, it takes ~10 half lives to fully get rid of it from the EMAs. With moving_average_half_life_ratio = 1. /. 80., that 2nd commit will only poluate 1 / 8 of the width of the smoothed curves.

val moving_average_relevance_threshold : float

See Exponential_moving_average