Module Irmin_traces.Trace_stat_summary_pp

Pretty printing of one or more summaries.

This file is NOT meant to be used from Tezos, as opposed to some other "trace_*" files.

This file contains A LOT of uninteresting boilerplate in order to build the pretty-printable table. Doing this using pandas-like multi-level dataframes would make the thing much more simpler.

module Utils = Trace_stat_summary_utils

module Summary = Trace_stat_summary

module Pb : sig ... end

val fprintf_result : Stdlib.Format.formatter -> string -> string -> string -> string -> string -> float -> unit

type summary = Summary.t

type scalar_format_fixed = [

| `SM

| `S3

| `Sm

| `Su

| `RG

| `RM

| `Ri

| `R3

| `P

]

Seconds minutes, Seconds 3 digits, Seconds milli, Seconds micro, Real giga, Real mega, Real as integer, Real 3 digits, Percent

val pp_scalar_fixed : Stdlib.Format.formatter -> ([< `P | `R3 | `RG | `RM | `Ri | `S3 | `SM | `Sm | `Su ] * float) -> unit

module Table0 : sig ... end

Summary

module Table1 : sig ... end

Highlights

module Table2 : sig ... end

module Table3 : sig ... end

module Table4 : sig ... end

Curves

val unsafe_pp : int -> Stdlib.Format.formatter -> string list -> Summary.t list -> unit

val pp : int -> Stdlib.Format.formatter -> (string list * Summary.t list) -> unit